Move test case to new file
